Apply code review fixes (1 critical, 3 major)
C1 — Stream-compare same-sized files instead of read_bytes()
 - core/fmt.py: add files_equal(a, b, chunk_size=1MB) helper that
   reads in chunks and short-circuits on first mismatch.
 - core/patch_builder.py and core/engines/dir_format.py call it in
   place of `src.read_bytes() == tgt.read_bytes()`. Was loading both
   copies fully into RAM (e.g. 38 GB for an unchanged 19 GB Unreal
   .ucas) — now bounded by the chunk size.

M1 — CLI --extra-file SRC[:DEST] is now drive-letter aware
 - cli/main.py: when entry begins with "X:", skip the first 2 chars
   when scanning for the SRC/DEST colon separator. Previously
   `C:\game\readme.txt:docs/readme.txt` parsed as src="C", dest=garbage
   on Windows.

M2 — PFMD container widened to uint64 data_len (version 2)
 - core/engines/dir_format.py: bump _VERSION to 2, write data_len as
   <Q (uint64) instead of <I (uint32). OP_NEW for files larger than
   4 GB no longer overflows.
 - stub/dir_patch_format.h: add pfmd_u64le, parser reads 8 bytes,
   uses additive bounds-check (data_len > end - p) to avoid pointer
   arithmetic overflow, requires version byte == PFMD_VERSION (2).
 - stub/xdelta3_stub.c, stub/jojodiff_stub.c: callback signatures
   take uint64_t data_len; apply_jojodiff() takes size_t patch_len.
 - Clean break — no v1 backward compat. All 12 prebuilt stubs
   rebuilt with MinGW-w64.

M3 — HDiffPatch directory args use os.sep
 - core/engines/hdiffpatch.py: rstrip("/\\") + os.sep instead of
   rstrip("/") + "/". Produces `C:\game\` on Windows and `/game/` on
   Linux. The trailing separator is what tells hdiffz to treat the
   arg as a directory.

21 tests passing under conda env xdelta.

+++ stub/xdelta3_stub.c
@@ -162,7 +162,7 @@ struct DirPatchCtx {
 };
 
 static int xd3_apply_entry(int op, const char *rel_path,
-                             const unsigned char *data, uint32_t data_len,
+                             const unsigned char *data, uint64_t data_len,
                              void *userdata)
 {
     struct DirPatchCtx *ctx = (struct DirPatchCtx *)userdata;
